requirements.txt
requests
beautifulsoup4
python-telegram-bot
pyrogram
tgcrypto
